fix: harden agent pipeline against device faults and concurrency
两轮代码审查的落地修复（12 项）：

- executor: 对外承诺「永不抛异常」。原实现未捕获 int(value) 的 ValueError，
  VLM 返回 wait:"一会儿" 这类脏数据会穿透主循环、让整个任务以 HTTP 500 中断；
  新增时长解析与 0-60s 范围钳制，SWIPE 支持全角逗号
- loop: 观察阶段（截图 / dump / focus）失败收敛为 ERROR 步骤并计入重试熔断，
  不再 500；任务进入 RUNNING 后任何未预期异常先落 FAILED 再抛，杜绝僵尸任务；
  post 观察失败降级为「未验证的 OK」而不是重发动作；Re-plan 路径统一走 VLM 验证，
  与成功路径判定标准一致；step 0 观察写入 history，/tasks/{id}/shots/0 可访问
- adb: screen_size 优先取 Override size（实际渲染尺寸，原取 Physical 会让
  归一化坐标整体偏移）；type_text 校验原始输入并拒绝 %；dump_ui 先清旧文件，
  并校验输出含 hierarchy 根节点
- grounding: 支持 2 个数字的点坐标；加守卫避免「微信 v8.0.32」被误判成坐标；
  像素坐标路径不再触发 screen_size 调用
- parser: bounds 支持负数（真实 UI 树有离屏负值）与单节点容错；匹配改为精确优先打分
- vlm: _extract_json 容错（VLM 常在 JSON 前后附带说明文字）；
  429/5xx/网络错误指数退避重试（4xx 不重试）；图片精度按阶段分级
- server: /tasks 默认后台线程执行，wait=true 时可同步等待；全局设备锁串行化
  写操作（409）；500 响应默认脱敏；非法 target 在契约层被拒（422），
  不再先拍截图碰真机再报参数错
- memory: 加 RLock 与锁内快照，避免并发读出自相矛盾的状态
- 新增 pyproject.toml，消除对 cwd 的依赖
- 新增 45 个回归测试用例，全部使用假设备，不需要 adb / 模拟器 / API Key

+++ models/action.py
@@ -19,8 +19,8 @@ class ActionType(str, Enum):
 
 
 class Point(BaseModel):
-    x: int
-    y: int
+    x: float
+    y: float
 
 
 class Action(BaseModel):
